Refactor CLI Structure
