+++ cbrain
@@ -4,13 +4,11 @@ CBRAIN CLI - Traditional shell executable
 """
 import sys
 import os
+from cbrain_cli.main import main
 
 # Add the current directory to Python path so we can import cbrain_cli.
 current_dir = os.path.dirname(os.path.abspath(__file__))
 sys.path.insert(0, current_dir)
-
-# Import and run the CLI.
-from cbrain_cli.main import main
 
 if __name__ == '__main__':
     sys.exit(main())
